requirements.txt
google-genai==1.10.0
flask==3.1.0
python-dotenv==1.1.0
Werkzeug==3.1.3
